Surrogate Synchrony - legacy

This is a legacy function to compute Surrogate Synchrony, please use susy function instead. Eventually legacy function may be deprecated in future.

Cross-correlations are computed up to a specific lag in seconds Maxlag, then aggregated within a chosen Segment of e.g. 30s. Aggregation is performed by transforming correlations to Fisher's Z, then computing mean Z in each segment, then across all segments of the time series. Segment shuffling is used to create surrogate time series, on which the same computations are run. This provides effect sizes ES. File: The pairs of time series are in the columns of the file, variable names are in the header line. If Automatic is clicked, the synchrony is computed of all adjacent pairs of columns in the file. If Automatic is unclicked, you may choose the two columns to be analyzed for synchrony, and two plots are additionally prepared. SUSY provides two different synchrony measures of each twin time series: mean Z and ES of mean Z; mean absolute_Z and ES of mean absolute_Z.

old.susy(
  filename="MEA-topstage-C3-Beethoven.txt", separator=" ", missings=".", datahead=TRUE,
  spalte1=2, spalte2=5, epoche=30, fps=15, maxlag=3*fps, automatic=FALSE,
  plotid=3, plotid2=5, pseudosVereinfachen=FALSE, anzahlPseudosGesamt=500,
  zufallsdaten=FALSE, anzahlzufallsdaten=10000, korrOhneBetragAnzeigen=TRUE,
  permutation=FALSE
)

Arguments

filename

Character scalar, name of a source file. Defaults to "MEA-topstage-C3-Beethoven.txt".

separator

Character scalar, source file data separator. Defaults to " ".

missings

Character scalar, source file data missings marker. Defalts to ".".

datahead

Logical scalar, tells if source data has a header row. Defaults to TRUE.

spalte1

Numeric, ???hier bezeichnete Spalte 1 wird mit Spalte 2 verglichen (falls automatic = FALSE)??. Defaults to 2.

spalte2

Numeric, ???. Defaults to 5.

epoche

Epochengroesse (SEGMENT SIZE) in s (Epoche darf nicht kleiner sein als Zeitfenster/fps). Epoche*fps darf nicht groesser sein als die Anzahl Daten (timepoints) geteilt durch 2. Beispiel 1: timepoints: 1000, fps = 10, also darf Epoche nicht groesser sein als 50s (=1000/(10*2)). Beispiel 2: maxlag : 90, fps = 10, also darf Epoche nicht KLEINER sein als 9s (=90/10).

fps

frames per second (SAMPLING RATE).

maxlag

Zeitfenster (WINDOW SIZE) fuer CCF - > maximaler lag in Sekunden (nur ganze Zahlen).

automatic

automatische Iteration? TRUE, FALSE. es wird die erste mit der zweiten Spalte verglichen, dann die dritte mit der vierten usw... ein Spaltenpaar muss dieselbe Anzahl Daten haben, aber verschiedene Paare nicht. Beachte: Es werden keine Plots ausgegeben falls automatic = TRUE.

plotid

welche Art plot soll angezeigt werden (man kann den plot auch manuell ausfuehren (unten)). 0 nichts, 1 GMcrosscorrs, 2 synchrony by epoch, 3 GM-Z, 4 Zeitreihenplot, 5 z not abs wenn plotid2 == 0, dann kein zweiter plot

plotid2

see above

pseudosVereinfachen

Berechnung der pseudo-Daten vereinfachen (TRUE, FALSE) Falls FALSE, anzahlPseudosGesamt irrelevant anzahlPseudos: Anzahl der Surrogats-tests aller Epochen

anzahlPseudosGesamt

...

zufallsdaten

Zufallsdaten (Normalverteilte Zahlen im Bereich [0,50]

anzahlzufallsdaten

...

korrOhneBetragAnzeigen

Korrelationen ohne Betrag auch anzeigen? (FALSE/TRUE)

permutation

Permuation: Synchronie f?r n(n-1)/2 Dyaden berechnen

Value

Returns NULL invisibly. A side effects include:

output to console, its form depends on korrOhneBetragAnzeigen argument.

plots, according to plotid1 and plotid2 arguments.

file "CrossCorrelations.txt" written to current working directory, if plotid1 or plotid2 specifies plot type 1.

See also

Examples

old.susy(zufallsdaten=TRUE, anzahlzufallsdaten=1000, plotid=0, plotid2=0)
#> Var1 Var2 n(data) Z Z-Pseudo SD(Z) SD(Z-Pseudo) n(lags) %>Pseudo n(Segmente) ES Z(lead1) Z(lead2) ES(lead1) ES(lead2) meanZ(in-phase) meanZ(anti-phase) Anzahl(in-phase) Anzahl(anti-phase) Z(noAbs) Z(Pseudo-noAbs) %>Pseudo(noAbs) ES(noAbs) 
#> randomA randomB 1000 0.03786327 0.03717355 0.01751604 0.02092609 91 49.45055 2 0.03295936 0.03697986 0.03854857 0.02771384 0.02529522 0.02796328 -0.02850865 51 40 0.003140455 0.002820419 52.74725 0.01012269