Surrogate Synchrony

Cross-correlations are computed up to a specific lag in seconds Maxlag, then aggregated within a chosen Segment of e.g. 30s. Aggregation is performed by transforming correlations to Fisher's Z, then computing mean Z in each segment, then across all segments of the time series. Segment shuffling is used to create surrogate time series, on which the same computations are run. This provides effect sizes ES. SUSY provides two different synchrony measures of each twin time series: mean Z and ES of mean Z; mean absolute_Z and ES of mean absolute_Z.
Legacy implementation is available at old.susy.

susy(x, permutation=FALSE, epoch=30L, fps=15L, maxlag=3L*fps,
  pseudo.simplify=FALSE, pseudo.total=500L)

Arguments

x

A data.frame of numeric columns.

permutation

Logical, default FALSE requires x to have even number of columns which are processed in pairs (1-2, 3-4, etc.). When permutation is TRUE then function computes all pairs combinations between columns provided in x (n*(n-1)/2 pairs).

epoch

Integer, segment size in seconds, default 30L. Must not be smaller than time window (fps), also epoch*fps must not be greater than the timepoints (nrow(x)/2). Example 1: timepoints: 1000, fps = 10, so the epoch must not be greater than 50s (=1000/(10*2)); Example 2: maxlag : 90, fps = 10, so epoch must not be less than 9s (=90/10).

fps

Integer, frames per second (sampling rate), default 15L.

maxlag

Integer, window size for ccf, maximum lag in seconds, default 3L*fps.

pseudo.simplify

Logical, default FALSE. Simplify calculation of the pseudo data.

pseudo.total

Integer, default 500L. Ignored when pseudo.simplify is FALSE (default).

Details

Segments are non-overlapping, and the number of segments that fit into the time series may have a remainder (usually a few seconds at the end of the time series), which is not considered.

Value

Object of class susy is returned. Each cross correlation pair is an element in resulting object.

See also

Examples

n = 1000
data = data.frame(
  var1 = runif(n,300,330),
  var2 = runif(n,300,330),
  var3 = runif(n,300,330)
)

## use only first two columns
res = susy(data[, 1:2])
length(res)
#> [1] 1
names(res)
#> [1] "var1-var2"

## use all columns and permutation
res = susy(data, permutation=TRUE)
length(res)
#> [1] 3
names(res)
#> [1] "var1-var2" "var1-var3" "var2-var3"

## print susy
res
#>   Var1 Var2 n(data)          Z   Z-Pseudo      SD(Z) SD(Z-Pseudo) n(lags)
#> 1 var1 var2    1000 0.03452312 0.03955354 0.01820813   0.01861211      91
#> 2 var1 var3    1000 0.03821652 0.03242691 0.02162886   0.01785053      91
#> 3 var2 var3    1000 0.03553745 0.03700488 0.01852510   0.01989256      91
#>   %>Pseudo n(Segmente)          ES   Z(lead1)   Z(lead2)    ES(lead1)
#> 1 36.26374           2 -0.27027679 0.03243375 0.03591506 -0.497470969
#> 2 60.43956           2  0.32433838 0.03581875 0.04031753 -0.002108751
#> 3 46.15385           2 -0.07376774 0.03607059 0.03522796 -0.032134130
#>    ES(lead2) meanZ(in-phase) meanZ(anti-phase) Anzahl(in-phase)
#> 1 -0.1381360      0.02439309       -0.02329122               47
#> 2  0.6946167      0.02912177       -0.02980047               51
#> 3 -0.1046060      0.02701247       -0.02279094               41
#>   Anzahl(anti-phase)      Z(noAbs) Z(Pseudo-noAbs) %>Pseudo(noAbs)   ES(noAbs)
#> 1                 44  0.0013369394   -0.0002166195        45.05495  0.04482044
#> 2                 40  0.0032218848    0.0024051151        43.95604  0.03109527
#> 3                 50 -0.0003520419    0.0031662224        47.25275 -0.10953553
print(res, legacy=TRUE)
#> Var1 Var2 n(data) Z Z-Pseudo SD(Z) SD(Z-Pseudo) n(lags) %>Pseudo n(Segmente) ES Z(lead1) Z(lead2) ES(lead1) ES(lead2) meanZ(in-phase) meanZ(anti-phase) Anzahl(in-phase) Anzahl(anti-phase) Z(noAbs) Z(Pseudo-noAbs) %>Pseudo(noAbs) ES(noAbs) 
#> var1 var2 1000 0.03452312 0.03955354 0.01820813 0.01861211 91 36.26374 2 -0.2702768 0.03243375 0.03591506 -0.497471 -0.138136 0.02439309 -0.02329122 47 44 0.001336939 -0.0002166195 45.05495 0.04482044 
#> var1 var3 1000 0.03821652 0.03242691 0.02162886 0.01785053 91 60.43956 2 0.3243384 0.03581875 0.04031753 -0.002108751 0.6946167 0.02912177 -0.02980047 51 40 0.003221885 0.002405115 43.95604 0.03109527 
#> var2 var3 1000 0.03553745 0.03700488 0.0185251 0.01989256 91 46.15385 2 -0.07376774 0.03607059 0.03522796 -0.03213413 -0.104606 0.02701247 -0.02279094 41 50 -0.0003520419 0.003166222 47.25275 -0.1095355 

## plot susy
plot(res)
plot(res, type=1:2)